requirements.txt
Django>=4.2
openai>=1.0.0
python-dotenv